Hurtling monsters/players tweaks and fixes.
This work all stems from NetHack 3.7 git commit 268022e and associated
commits made in regards to it. The devteam (paxed specifically) had
added a new wizmode command (#wiztelekinesis) that allows one to test
hurtling monsters and examine the results. This was done in conjunction
with some fixes to the monster hurtling code. qt and I took this a step
further, allowing #wiztelekinesis to hurtle the player as well, and qt
provided a fix for how hurtling over various terrain behaves. I've also
added some code here that takes air terrain in Vlad's cavern into
account.

There may be future tweaks and fixes to this. And this opens the door
for some interesting behavior (e.g. force bolt spell at a high enough
skill level could push its target back).

+++ include/extern.h
@@ -1251,7 +1251,7 @@ E void FDECL(newmonhp, (struct monst *, int));
 E struct mextra *NDECL(newmextra);
 E void FDECL(copy_mextra, (struct monst *, struct monst *));
 E void FDECL(dealloc_mextra, (struct monst *));
-E struct monst *FDECL(makemon, (struct permonst *, int, int, int));
+E struct monst *FDECL(makemon, (struct permonst *, int, int, long));
 E boolean FDECL(create_critters, (int, struct permonst *, BOOLEAN_P));
 E struct permonst *NDECL(rndmonst);
 E void FDECL(reset_rndmonst, (int));
@@ -2591,7 +2591,7 @@ E boolean FDECL(stucksteed, (BOOLEAN_P));
 
 /* ### teleport.c ### */
 
-E boolean FDECL(goodpos, (int, int, struct monst *, unsigned));
+E boolean FDECL(goodpos, (int, int, struct monst *, long));
 E boolean FDECL(enexto, (coord *, XCHAR_P, XCHAR_P, struct permonst *));
 E boolean FDECL(enexto_core, (coord *, XCHAR_P, XCHAR_P,
                               struct permonst *, unsigned));
